new motor and new structure

+++ Action_User/pose_plan.h
@@ -9,11 +9,11 @@
 
 
 
-void BaseCenterUpdate(const float ChainAngle[3],pcoordinate_t BaseCenter_t);
-void ChainUpdateByBase(const coordinate_t BaseCenter_t, float ChainAngle[3]);
-void AngularVelUpdate(const coordinate_t BaseCenter_t,const float ChainAngle[3],const float Velocity[3],float AngularVel[3]);
-void VelocityUpdate(const coordinate_t BaseCenter_t,const float ChainAngle[3],const float AngularVel[3],float Velocity[3]);
-float getK(coordinate_t startPoint,coordinate_t endPoint);
+void BaseCenterUpdate(const double ChainAngle[3],pcoordinate_t BaseCenter_t);
+void ChainUpdateByBase(const coordinate_t BaseCenter_t, double ChainAngle[3]);
+void AngularVelUpdate(const coordinate_t BaseCenter_t,const double ChainAngle[3],const double Velocity[3],double AngularVel[3]);
+void VelocityUpdate(const coordinate_t BaseCenter_t,const double ChainAngle[3],const double AngularVel[3],double Velocity[3]);
+double getK(coordinate_t startPoint,coordinate_t endPoint);
 #endif
